fixed: store the control values as double for turning into JSON, and as float for passing to the LADSPA plugin, using class controlvalue to keep them synchronized, since javascript uses double as well, this will make the double->string->double conversions identical

+++ src/jackclient.cpp
@@ -222,7 +222,7 @@ void CJackClient::InitLadspa()
   }
 }
 
-void CJackClient::UpdateGain(float gain, int index)
+void CJackClient::UpdateGain(double gain, int index)
 {
   //update the gain, this will be read from the jack thread
   CLock lock(m_mutex);
